refactor: move spinner and geolocation button
