--- site/src/components/geolocation_button/index.ts
@@ -1,1 +1,0 @@
-export * from './geolocation_button'
