+++ site/src/features/pages/single_train/components/page.tsx
@@ -11,6 +11,7 @@ import {
   itemIcon,
 } from '@junat/ui/components/dropdown_menu/index'
 import { ErrorMessageWithRetry } from '@junat/ui/components/error_message'
+import { Spinner } from '@junat/ui/components/spinner'
 import { useToast } from '@junat/ui/components/toast/index'
 import Calendar from '@junat/ui/icons/calendar.svg'
 import CirclesHorizontal from '@junat/ui/icons/circles_horizontal.svg'
@@ -18,7 +19,6 @@ import ObjectHorizontalLeft from '@junat/ui/icons/object_horizontal_left.svg'
 import Share from '@junat/ui/icons/share.svg'
 
 import { Head } from '~/components/head'
-import { Spinner } from '~/components/spinner'
 import { translate, useLocale, useTranslations } from '~/i18n'
 import Page from '~/layouts/page'
 import { getNewTrainPath, getTrainTitle, handleShare } from '../helpers'
